refactor passoff test grader
