+++ src/test/java/edu/byu/cs/autograder/score/days-late-with-holidays-friday-holiday-and-consecutive-holidays.jpg
